Compressed linear system solver for FDM solvers (#124)
This revision lets CSR matrix to be used for FDM solvers. This approach can save computational time when the fluid is occupying computational domain sparsely. Roughly, if the fluid volume is 25% of the entire domain, it can perform 2x faster than the grid-based matrix. Note that the existing FdmLinearSystem2 and 3 are also sparse matrices; The new linear system is a sparse matrix AND compressed by adapting to the fluid volume. We still leave this new linear system to be an option which is not turned on by default since it can consume more memory and time for typical situations and also not suitable for smoke simulations.